basic tools for performance/stress testing
push_nginx.py is a utility for sending a lot of push messages to the server in parallel (parallelized with processes).
It's highly configurable from the command line and the emphasis is on testing configurations that have many channels.

simple_client.py is a utility for emulating a small number of clients listening on these channels.